# contracts/openapi/aihub-v1.yaml
...
openapi: 3.1.0
...
info:
  title: AIHub Platform API
  version: v1
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    内部 AI 资产管理平台 REST API 契约。统一成功响应为 ApiResponse，统一失败响应为 ApiError。
    Agent 不应依赖本地化 message 做分支判断，只依赖稳定的 code 与 retryable。
servers:
  - url: /api/v1
    description: 统一 API 前缀
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      operationId: login
      summary: 本地用户登录并签发 JWT
      tags: [auth]
      security: []
      x-implementation-status: planned-p0-b
      x-audit-event: AUTH_LOGIN_SUCCEEDED_OR_FAILED
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 登录成功；浏览器刷新 JWT 同时写入 Secure/HttpOnly/SameSite Cookie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '423': { $ref: '#/components/responses/ErrorResponse' }
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: 轮换刷新 JWT 并签发新访问 JWT
      tags: [auth]
      security: []
      x-implementation-status: planned-p0-b
      x-audit-event: AUTH_TOKEN_REFRESHED_OR_REPLAY_REJECTED
      requestBody:
        required: false
        description: 浏览器使用安全 Cookie；非浏览器客户端可在请求体提交刷新 JWT。
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: 刷新成功，旧刷新 JWT 立即失效
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /auth/token:
    post:
      operationId: exchangeClientCredential
      summary: Agent/Service/API Client 凭据换取 JWT
      tags: [auth]
      security: []
      x-implementation-status: planned-p0-b
      x-audit-event: CLIENT_TOKEN_ISSUED_OR_REJECTED
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientCredentialTokenRequest' }
      responses:
        '200':
          description: 凭据有效，签发访问与轮换刷新 JWT
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
  /auth/logout:
    post:
      operationId: logout
      summary: 吊销当前 Token Family
      tags: [auth]
      x-implementation-status: planned-p0-b
      x-audit-event: AUTH_LOGOUT
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /me:
    get:
      operationId: getCurrentPrincipal
      summary: 获取当前主体、角色和粗粒度 Scope
      tags: [auth]
      x-implementation-status: planned-p0-b
      responses:
        '200':
          description: 当前主体
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CurrentPrincipal' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /me/password:
    put:
      operationId: changeCurrentUserPassword
      summary: 修改当前本地用户密码并吊销其他 Token Family
      tags: [auth]
      x-implementation-status: planned-p0-b
      x-audit-event: USER_PASSWORD_CHANGED
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200':
          description: 密码修改成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
...
  /system/organizations:
    get:
      operationId: listOrganizations
      summary: 查询调用者可管理或可访问的组织
      tags: [organizations]
      x-implementation-status: planned-p0-b
      x-required-permissions: [project:view]
      responses:
        '200':
          description: 组织列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/OrganizationView' } }
        '403': { $ref: '#/components/responses/ErrorResponse' }
    post:
      operationId: createOrganization
      summary: 创建组织作用域
      tags: [organizations]
      x-implementation-status: planned-p0-b
      x-required-permissions: [organization:manage]
      x-audit-event: ORGANIZATION_CREATED
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationRequest' }
      responses:
        '201':
          description: 组织创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/OrganizationView' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
...
  /system/organizations/{organizationId}/projects:
    parameters:
      - { name: organizationId, in: path, required: true, schema: { type: string, pattern: '^org_' } }
    get:
      operationId: listOrganizationProjects
      summary: 查询组织项目
      tags: [organizations]
      x-implementation-status: planned-p0-b
      x-required-permissions: [project:view]
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/ProjectView' } }
        '403': { $ref: '#/components/responses/ErrorResponse' }
    post:
      operationId: createProject
      summary: 创建组织项目
      tags: [organizations]
      x-implementation-status: planned-p0-b
      x-required-permissions: [project:manage]
      x-audit-event: PROJECT_CREATED
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectRequest' }
      responses:
        '201':
          description: 项目创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ProjectView' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
...
  /system/resource-acls:
    get:
      operationId: listResourceAcls
      summary: 查询资源 ACL
      tags: [authorization]
      x-implementation-status: planned-p0-b
      x-required-permissions: [authorization:read]
      responses:
        '200':
          description: ACL 列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/ResourceAclView' } }
        '403': { $ref: '#/components/responses/ErrorResponse' }
    post:
      operationId: createResourceAcl
      summary: 创建资源 ACL
      tags: [authorization]
      x-implementation-status: planned-p0-b
      x-required-permissions: [authorization:manage]
      x-audit-event: RESOURCE_ACL_CREATED
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateResourceAclRequest' }
      responses:
        '201':
          description: ACL 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ResourceAclView' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /system/resource-acls/{aclId}:
    delete:
      operationId: deleteResourceAcl
      summary: 删除资源 ACL
      tags: [authorization]
      x-implementation-status: planned-p0-b
      x-required-permissions: [authorization:manage]
      x-audit-event: RESOURCE_ACL_DELETED
      parameters:
        - { name: aclId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 已删除, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        平台签发的短期访问 JWT。JWT 使用非对称签名和 kid；资源级权限由后端实时判断，
        不以 JWT 内角色替代 RBAC/ACL/资源状态校验。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 当前 Principal 范围内唯一的写请求幂等键。
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    ErrorResponse:
      description: 统一失败响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiResponse:
      type: object
      description: 统一成功响应包装
      required:
        - requestId
        - traceId
        - timestamp
      properties:
        data:
          description: 业务数据载荷，结构由具体接口定义
          type: [object, array, string, number, integer, boolean, 'null']
        requestId:
          type: string
          description: 请求标识，回显 X-Request-Id
          example: req_01J0000000000000000000000
        traceId:
          type: string
          description: 分布式追踪标识
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        timestamp:
          type: string
          format: date-time
          description: 服务端 UTC 时间戳
          example: '2026-06-29T08:00:00Z'
    ApiError:
      type: object
      description: 统一失败响应
      required:
        - code
        - message
        - retryable
        - requestId
        - traceId
      properties:
        code:
          type: string
          description: 稳定的语义错误码，分支判断只依赖此字段
          example: ASSET_VERSION_CONFLICT
        message:
          type: string
          description: 默认（可本地化）错误消息，不用于分支判断
          example: Version already exists
        i18nKey:
          type: string
          description: 国际化文案 Key
          example: error.asset.versionConflict
        details:
          type: object
          description: 附加上下文，结构由错误码定义
          additionalProperties: true
        retryable:
          type: boolean
          description: 是否可重试；仅当为 true 时 Agent 才应按退避策略重试
          example: false
        requestId:
          type: string
          example: req_01J0000000000000000000000
        traceId:
          type: string
          example: 4bf92f3577b34da6a3ce929d0e0e4736
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          format: password
          writeOnly: true
          minLength: 12
          maxLength: 128
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      properties:
        refreshToken:
          type: string
          format: password
          writeOnly: true
          description: 非浏览器客户端提交；浏览器优先使用 Secure/HttpOnly/SameSite Cookie。
    ClientCredentialTokenRequest:
      type: object
      additionalProperties: false
      required: [grantType, subjectId, credential]
      properties:
        grantType:
          type: string
          const: client_credentials
        subjectId:
          type: string
          description: Agent、Service 或 API Client 的稳定业务 ID。
        credential:
          type: string
          format: password
          writeOnly: true
    TokenPair:
      type: object
      required: [tokenType, accessToken, expiresIn, refreshExpiresIn, principal]
      properties:
        tokenType:
          type: string
          const: Bearer
        accessToken:
          type: string
          description: 默认 15 分钟有效；客户端不得写入持久化浏览器存储。
        refreshToken:
          type: string
          description: 默认 7 天有效并在每次刷新时轮换；浏览器响应可仅通过安全 Cookie 返回。
        expiresIn:
          type: integer
          description: 访问 JWT 剩余秒数。
          example: 900
        refreshExpiresIn:
          type: integer
          description: 刷新 JWT 剩余秒数。
          example: 604800
        principal:
          $ref: '#/components/schemas/CurrentPrincipal'
    PrincipalType:
      type: string
      enum: [USER, AGENT, SERVICE, API_CLIENT, WORKER]
    CurrentPrincipal:
      type: object
      required: [principalId, principalType, subject, roles, scopes, locale]
      properties:
        principalId: { type: string, pattern: '^prn_' }
        userId:
          type: [string, 'null']
          pattern: '^usr_'
        principalType: { $ref: '#/components/schemas/PrincipalType' }
        subject: { type: string }
        displayName: { type: [string, 'null'] }
        organizationId: { type: [string, 'null'] }
        roles:
          type: array
          items: { type: string }
        scopes:
          type: array
          items: { type: string }
        locale: { type: string, example: zh-CN }
        forcePasswordChange: { type: boolean, default: false }
    PrincipalStatus:
      type: string
      enum: [ACTIVE, LOCKED, DISABLED]
...
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          writeOnly: true
        newPassword:
          type: string
          format: password
          writeOnly: true
          minLength: 12
          maxLength: 128
...
    CreateOrganizationRequest:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$'
        name: { type: string, minLength: 1, maxLength: 128 }
        giteaOrganization: { type: [string, 'null'], maxLength: 128 }
    OrganizationView:
      type: object
      required: [organizationId, code, name, status, createdAt]
      properties:
        organizationId: { type: string, pattern: '^org_' }
        code: { type: string }
        name: { type: string }
        giteaOrganization: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/PrincipalStatus' }
        createdAt: { type: string, format: date-time }
...
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$'
        name: { type: string, minLength: 1, maxLength: 128 }
    ProjectView:
      type: object
      required: [projectId, organizationId, code, name, status, createdAt]
      properties:
        projectId: { type: string, pattern: '^prj_' }
        organizationId: { type: string, pattern: '^org_' }
        code: { type: string }
        name: { type: string }
        status: { $ref: '#/components/schemas/PrincipalStatus' }
        createdAt: { type: string, format: date-time }
...
    CreateResourceAclRequest:
      type: object
      additionalProperties: false
      required: [principalId, resourceType, resourceId, permissionCodes]
      properties:
        principalId: { type: string, pattern: '^prn_' }
        resourceType: { type: string, example: ASSET }
        resourceId: { type: string }
        permissionCodes:
          type: array
          uniqueItems: true
          minItems: 1
          items: { type: string }
    ResourceAclView:
      type: object
      required: [aclId, principalId, resourceType, resourceId, permissionCodes, createdAt]
      properties:
        aclId: { type: string }
        principalId: { type: string }
        resourceType: { type: string }
        resourceId: { type: string }
        permissionCodes:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...